progress on 11/9/2023 --
simplified some of the code, at least in terms of how it can be seen visually. 3 header files have been created to be in charge of different components. after speaking with jensen, it looks like creating classes would be similar and a more practical method of what im doing now. additionally, the encoders were figured out and i learned a little bit about interrupts, which we will need 3 interrupt pins in our robot for the encoders. that's a test file that hasnt been uploaded yet